feat(core): 提供新的密码配置形式
